aiogram_v2_insta_bot/requirements.txt
aiogram
aiohttp
instaloader
python-dotenv
